Move getReserve JS implementation into getReserves unit test
